refactor so code is not repeated
